+++ Changelog.md
@@ -3,6 +3,7 @@
 ### HEAD
 --------------
 
+- Fix `Router` to handle same url with different HTTP methods
 
 ### 1.0.0
 ------------